requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
Django==5.0.7
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
psycopg2-binary==2.9.9
PyPDF2==3.0.1
PySocks==1.7.1
python-decouple==3.8
pyunitreport==0.1.4
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
